Add home.js
