+++ old/js/all.js
@@ -1,36 +1,3 @@
-for (let e of document.querySelectorAll(".close")) {
-    e.onclick = modalClose;
-}
-
-window.onclick = function (event) {
-    if (modals.find(x => x.element == event.target)) {
-        modalClose(event.target);
-    }
-};
-
-function shouldProcessMutations(mutationList) {
-    let processThis = false;
-
-    // ensure we're processing more than an addition of something this very handler added
-    for (let mutation of mutationList) {
-        for (let addedElem of mutation.addedNodes) {
-            if (
-                addedElem.classList &&
-                !addedElem.classList.contains("splus-addedtodynamicdropdown")
-            ) {
-                processThis = true;
-                break;
-            }
-        }
-
-        if (processThis) {
-            break;
-        }
-    }
-
-    return processThis;
-}
-
 let siteNavigationTileHelpers = {
     createSpacerTile: function () {
         return createElement("div", [
@@ -554,509 +521,5 @@ let siteNavigationTileHelpers = {
 
     docObserver.observe(document.body, { childList: true, subtree: true });
 })();
-
-async function createQuickAccess() {
-    let rightCol = document.getElementById("right-column-inner");
-    let linkWrap;
-
-    let wrapper = createElement("div", ["quick-access-wrapper"], {}, [
-        createElement("h3", ["h3-med"], { title: "Added by Schoology Plus" }, [
-            createSvgLogo("splus-logo-inline"),
-            // createElement("img", ["splus-logo-inline"], { src: chrome.runtime.getURL("imgs/plus-icon.png"), title: "Provided by Schoology Plus" }),
-            createElement("span", [], { textContent: "Quick Access" }),
-            createElement("a", ["quick-right-link", "splus-track-clicks"], {
-                id: "quick-access-splus-settings",
-                textContent: "Customize Sidebar",
-                href: "#splus-settings#setting-input-sidebarSectionOrder",
-            }),
-        ]),
-        createElement("div", ["date-header", "first"], {}, [
-            createElement("h4", [], { textContent: "Pages" }),
-        ]),
-        (linkWrap = createElement("div", ["quick-link-wrapper"])),
-    ]);
-
-    const PAGES = [
-        { textContent: "Grade Report", href: "/grades/grades", id: "quick-access-grades" },
-        { textContent: "Courses", href: "/courses", id: "quick-access-courses" },
-        { textContent: "Mastery", href: "/mastery", id: "quick-access-mastery" },
-        { textContent: "Groups", href: "/groups", id: "quick-access-groups" },
-        { textContent: "Messages", href: "/messages", id: "quick-access-messages" },
-    ];
-
-    for (let page of PAGES) {
-        let a = linkWrap.appendChild(
-            createElement("a", ["quick-link", "splus-track-clicks"], page)
-        );
-        a.dataset.splusTrackingContext = "Quick Access";
-    }
-
-    wrapper.appendChild(
-        createElement("div", ["date-header"], {}, [
-            createElement("h4", [], {}, [
-                createElement("span", [], { textContent: "Courses" }),
-                createElement("a", ["quick-right-link", "splus-track-clicks"], {
-                    id: "quick-access-reorder",
-                    textContent: "Reorder",
-                    href: "/courses?reorder",
-                }),
-            ]),
-        ])
-    );
-
-    try {
-        let sectionsList = (await fetchApiJson(`users/${getUserId()}/sections`)).section;
-
-        if (!sectionsList || sectionsList.length == 0) {
-            wrapper.appendChild(
-                createElement("p", ["quick-access-no-courses"], { textContent: "No courses found" })
-            );
-        } else {
-            let courseOptionsButton;
-            let iconImage;
-            let courseIconsContainer;
-            for (let section of sectionsList) {
-                wrapper.appendChild(
-                    createElement("div", ["quick-access-course"], {}, [
-                        (iconImage = createElement("div", ["splus-course-icon"], {
-                            dataset: {
-                                courseTitle: `${section.course_title}: ${section.section_title}`,
-                            },
-                        })),
-                        createElement("a", ["splus-track-clicks", "quick-course-link"], {
-                            textContent: `${section.course_title}: ${section.section_title}`,
-                            href: `/course/${section.id}`,
-                            dataset: {
-                                splusTrackingId: "quick-access-course-link",
-                                splusTrackingContext: "Quick Access",
-                            },
-                        }),
-                        (courseIconsContainer = createElement("div", ["icons-container"], {}, [
-                            createElement("a", ["icon", "icon-grades", "splus-track-clicks"], {
-                                href: `/course/${section.id}/student_grades`,
-                                title: "Grades",
-                                dataset: {
-                                    splusTrackingId: "quick-access-grades-link",
-                                    splusTrackingContext: "Quick Access",
-                                },
-                            }),
-                            createElement("a", ["icon", "icon-mastery", "splus-track-clicks"], {
-                                href: `/course/${section.id}/student_mastery`,
-                                title: "Mastery",
-                                dataset: {
-                                    splusTrackingId: "quick-access-mastery-link",
-                                    splusTrackingContext: "Quick Access",
-                                },
-                            }),
-                            (courseOptionsButton = createElement(
-                                "a",
-                                ["icon", "icon-settings", "splus-track-clicks"],
-                                {
-                                    href: "#",
-                                    dataset: {
-                                        splusTrackingId: "quick-access-settings-link",
-                                        splusTrackingContext: "Quick Access",
-                                    },
-                                }
-                            )),
-                        ])),
-                    ])
-                );
-
-                let quickLink = Setting.getNestedValue("courseQuickLinks", section.id);
-                if (quickLink && quickLink !== "") {
-                    courseIconsContainer.prepend(
-                        createElement("a", ["icon", "icon-quicklink", "splus-track-clicks"], {
-                            href: quickLink,
-                            title: `Quick Link \n(${quickLink})`,
-                            dataset: {
-                                splusTrackingId: "quick-access-quicklink-link",
-                                splusTrackingContext: "Quick Access",
-                            },
-                        })
-                    );
-                }
-
-                iconImage.style.backgroundImage = `url(${chrome.runtime.getURL(
-                    "imgs/fallback-course-icon.svg"
-                )})`;
-
-                courseOptionsButton.addEventListener("click", () =>
-                    openModal("course-settings-modal", {
-                        courseId: section.id,
-                        courseName: `${section.course_title}: ${section.section_title}`,
-                    })
-                );
-            }
-        }
-    } catch (err) {
-        if (err === "noapikey") {
-            wrapper.appendChild(
-                createElement("div", ["quick-access-no-api"], {}, [
-                    createElement("p", [], {
-                        textContent:
-                            "Please grant access to your enrolled courses in order to use this feature.",
-                    }),
-                    createButton("quick-access-grant-access", "Grant Access", () => {
-                        location.pathname = "/api";
-                    }),
-                ])
-            );
-        } else {
-            throw err;
-        }
-    }
-
-    rightCol.append(wrapper);
-}
-
-function getAssignmentId(url) {
-    if (url.includes("/assignment/")) {
-        return url.match(/assignment\/(\d+)/)[1];
-    } else if (url.includes("/course/")) {
-        // Discussion boards, maybe other assignments as well
-        return url.match(/course\/\d+\/.*\/(\d+)/)[1];
-    } else if (url.includes("/event/")) {
-        // Calendar events
-        return url.match(/event\/(\d+)/)[1];
-    } else if (url.includes("/external_tool/")) {
-        // External tools
-        return url.match(/external_tool\/(\d+)/)[1];
-    }
-
-    return null;
-}
-
-function indicateSubmittedAssignments() {
-    let upcomingList = document.querySelector(".upcoming-events .upcoming-list");
-    const completionOverridesSetting = "assignmentCompletionOverrides";
-    const assignCompleteClass = "splus-assignment-complete";
-    const assignIncompleteClass = "splus-assignment-notcomplete";
-
-    // checks on the backend if an assignment is complete (submitted)
-    // does not check user overrides
-    async function isAssignmentCompleteAsync(assignmentId) {
-        if (assignmentId == null) {
-            return false;
-        }
-        try {
-            let revisionData = await fetchApiJson(`dropbox/${assignmentId}/${getUserId()}`);
-            let revisions = revisionData.revision;
-
-            return !!(revisions && revisions.length && !revisions[revisions.length - 1].draft);
-        } catch (err) {
-            Logger.warn(
-                `Couldn't determine if assignment ${assignmentId} was complete. This is likely not a normal assignment.`
-            );
-            return false;
-        }
-    }
-
-    // checks user override for assignment completion
-    function isAssignmentMarkedComplete(assignmentId) {
-        return !!Setting.getNestedValue(completionOverridesSetting, assignmentId);
-    }
-
-    function setAssignmentCompleteOverride(assignmentId, isComplete) {
-        isComplete = !!isComplete;
-
-        let overrides = Setting.getValue(completionOverridesSetting);
-
-        if (!overrides && !isComplete) return;
-        else if (!overrides) overrides = {};
-
-        if (!isComplete) {
-            delete overrides[assignmentId];
-        } else {
-            overrides[assignmentId] = isComplete;
-        }
-
-        Setting.setValue(completionOverridesSetting, overrides);
-    }
-
-    function createAssignmentSubmittedCheckmarkIndicator(eventElement, assignmentId) {
-        let elem = document.createElement("button");
-        elem.classList.add("splus-completed-check-indicator");
-        elem.addEventListener("click", function () {
-            // if we're "faux-complete" and clicked, unmark the forced state
-            if (
-                eventElement.classList.contains(assignCompleteClass) &&
-                isAssignmentMarkedComplete(assignmentId)
-            ) {
-                eventElement.classList.remove(assignCompleteClass);
-                setAssignmentCompleteOverride(assignmentId, false);
-                trackEvent("button_click", {
-                    id: "splus-completed-check-indicator",
-                    context: "Checklist",
-                    value: "uncheck",
-                    legacyTarget: "splus-completed-check-indicator",
-                    legacyAction: "uncheck",
-                    legacyLabel: "Checkmarks",
-                });
-                // TODO handle async nicely
-                processAssignmentUpcomingAsync(eventElement);
-                // if we're incomplete and click, force the completed state
-            } else if (eventElement.classList.contains(assignIncompleteClass)) {
-                eventElement.classList.remove(assignIncompleteClass);
-                trackEvent("button_click", {
-                    id: "splus-completed-check-indicator",
-                    context: "Checklist",
-                    value: "check",
-                    legacyTarget: "splus-completed-check-indicator",
-                    legacyAction: "check",
-                    legacyLabel: "Checkmarks",
-                });
-                setAssignmentCompleteOverride(assignmentId, true);
-                // TODO handle async nicely
-                processAssignmentUpcomingAsync(eventElement);
-            }
-        });
-        return elem;
-    }
-
-    // returns assignment ID for convenience
-    async function processAssignmentUpcomingAsync(eventElement) {
-        let infotipElement = eventElement.querySelector(".infotip, .singleday");
-        let assignmentElement = infotipElement.querySelector("a[href]");
-
-        // TODO errorcheck the assignmentId match
-        let assignmentId = getAssignmentId(assignmentElement.href);
-
-        // add a CSS class for both states, so we can distinguish 'loading' from known-(in)complete
-        let isMarkedComplete = isAssignmentMarkedComplete(assignmentId);
-        if (isMarkedComplete || (await isAssignmentCompleteAsync(assignmentId))) {
-            Logger.log(
-                `Marking assignment ${assignmentId} as complete ✔ (is force-marked complete? ${isMarkedComplete})`
-            );
-            eventElement.classList.add(assignCompleteClass);
-        } else {
-            eventElement.classList.add(assignIncompleteClass);
-            Logger.log(`Assignment ${assignmentId} is not submitted`);
-        }
-
-        if (!eventElement.querySelector(".splus-completed-check-indicator")) {
-            infotipElement.insertAdjacentElement(
-                infotipElement.classList.contains("singleday") ? "afterbegin" : "afterend",
-                createAssignmentSubmittedCheckmarkIndicator(eventElement, assignmentId)
-            );
-        }
-
-        return assignmentId;
-    }
-
-    // Indicate submitted assignments in Upcoming
-    async function indicateSubmitted() {
-        Logger.log("Checking to see if upcoming assignments are submitted");
-        let idSet = new Set();
-        for (let upcomingList of document.querySelectorAll(".upcoming-list")) {
-            switch (Setting.getValue("indicateSubmission")) {
-                case "disabled":
-                    break;
-                case "strikethrough":
-                    upcomingList.classList.add("splus-mark-completed-strikethrough");
-                    break;
-                case "hide":
-                    upcomingList.classList.add("splus-mark-completed-hide");
-                    break;
-                case "check":
-                default:
-                    upcomingList.classList.add("splus-mark-completed-check");
-                    break;
-            }
-
-            let upcomingEventElements = upcomingList.querySelectorAll(
-                ".upcoming-event:not(.upcoming-subevents-block)"
-            );
-
-            for (let eventElement of upcomingEventElements) {
-                try {
-                    idSet.add(await processAssignmentUpcomingAsync(eventElement));
-                } catch (err) {
-                    Logger.error(
-                        `Failed checking assignment '${
-                            eventElement.querySelector(".infotip a[href]")?.href
-                        }' : `,
-                        err
-                    );
-                }
-            }
-        }
-
-        // check if reload is present and visible on page
-        let reloadButton = upcomingList.querySelector("button.button-reset.refresh-button");
-        if (reloadButton && reloadButton.offsetParent !== null) {
-            reloadButton.addEventListener("click", () =>
-                setTimeout(() => {
-                    indicateSubmitted();
-
-                    try {
-                        document.getElementById("todo")?.remove();
-                        let overdueHeading = document.querySelector(
-                            `${SIDEBAR_SECTIONS_MAP["Overdue"].selector} h4`
-                        );
-                        overdueHeading?.replaceWith(
-                            createElement("h3", [], {
-                                style: { textTransform: "capitalize" },
-                                textContent: overdueHeading.textContent.toLowerCase(),
-                            })
-                        );
-                        let upcomingHeading = document.querySelector(
-                            `${SIDEBAR_SECTIONS_MAP["Upcoming"].selector} h4`
-                        );
-                        upcomingHeading?.replaceWith(
-                            createElement("h3", [], {
-                                style: { textTransform: "capitalize" },
-                                textContent: upcomingHeading.textContent.toLowerCase(),
-                            })
-                        );
-                    } catch {}
-                }, 500)
-            );
-        } else {
-            // loaded properly
-            // clear out old assignments from local cache which aren't relevant anymore
-            let overrides = Setting.getValue(completionOverridesSetting);
-
-            if (overrides) {
-                for (var key in overrides) {
-                    if (overrides.hasOwnProperty(key) && !idSet.has(key)) {
-                        delete overrides[key];
-                    }
-                }
-                Setting.setValue(completionOverridesSetting, overrides);
-                Logger.info("Done clearing old overrides");
-            }
-        }
-    }
-
-    setTimeout(indicateSubmitted, 1000);
-}
-
-function getRecentlyCompletedDenominators() {
-    let recentlyCompletedList = document.querySelector(
-        ".recently-completed-wrapper .recently-completed-list"
-    );
-
-    async function getDirectAssignmentDenominatorAsync(assignmentId) {
-        try {
-            let json = await fetchApiJson(`sections/${sectionId}/assignments/${assignmentId}`);
-            return json.max_points;
-        } catch (err) {
-            return null;
-        }
-    }
-
-    async function getAssignmentDenominatorAsync(sectionId, assignmentId) {
-        if (assignmentId == null) {
-            return null;
-        }
-
-        let directDenominator = await getDirectAssignmentDenominatorAsync(assignmentId);
-        if (directDenominator !== null && !Number.isNaN(directDenominator)) {
-            Logger.debug(
-                `Found direct denominator for assignment ${assignmentId} in section ${sectionId}: ${directDenominator}`
-            );
-            return directDenominator;
-        }
-
-        try {
-            let json = await fetchApiJson(`users/${getUserId()}/grades?section_id=${sectionId}`);
-
-            if (json.section.length === 0) {
-                throw new Error("Assignment details could not be read");
-            }
-
-            const assignments = json.section[0].period.reduce(
-                (prevVal, curVal) => prevVal.concat(curVal.assignment),
-                []
-            ); //combines the assignment arrays from each period
-
-            let denom = Number.parseFloat(
-                assignments.filter(x => x.assignment_id == assignmentId)[0].max_points
-            );
-
-            Logger.debug(
-                `Found indirect denominator for assignment ${assignmentId} in section ${sectionId}: ${denom}`
-            );
-
-            return denom;
-        } catch (err) {
-            Logger.error(
-                `Failed finding denominator for assignment ${assignmentId} in section ${sectionId}`,
-                err
-            );
-            return null;
-        }
-    }
-
-    async function getSectionIdMap() {
-        let sections = await fetchApiJson(`users/${getUserId()}/sections`);
-        let sectionMap = {};
-
-        for (let section of sections.section) {
-            sectionMap[section.course_title + " : " + section.section_title] = section.id;
-        }
-
-        return sectionMap;
-    }
-
-    // Indicate submitted assignments in Upcoming
-    async function getDenominators() {
-        let sectionMap = await getSectionIdMap();
-
-        for (let recentEvent of recentlyCompletedList.querySelectorAll(
-            ".recently-completed-event"
-        )) {
-            try {
-                let eventLink = recentEvent.querySelector("a[href]");
-                let assignmentId = getAssignmentId(eventLink.href);
-                let sectionId =
-                    sectionMap[
-                        recentEvent
-                            .querySelector(".realm-title-course-title .realm-main-titles")
-                            .textContent.trim()
-                    ];
-
-                if (sectionId && assignmentId) {
-                    Logger.debug(
-                        `Getting denominator for assignment ${assignmentId} in section ${sectionId}`
-                    );
-                    let denominator = await getAssignmentDenominatorAsync(sectionId, assignmentId);
-                    Logger.debug(
-                        `Got denominator for assignment ${assignmentId} in section ${sectionId}: ${denominator}`
-                    );
-
-                    if (denominator) {
-                        let prevElement = recentEvent.querySelector(
-                            "span.infotip.grade-infotip span.recently-completed-grade"
-                        );
-
-                        if (prevElement) {
-                            let denominatorElement = createElement(
-                                "span",
-                                ["splus-recent-denominator"],
-                                { textContent: ` / ${denominator}` }
-                            );
-                            prevElement.insertAdjacentElement("afterend", denominatorElement);
-                        } else {
-                            recentEvent.querySelector(
-                                "span.recently-completed-grade"
-                            ).textContent += ` / ${denominator}`;
-                        }
-                    }
-                }
-            } catch (err) {
-                Logger.error(
-                    `Failed finding denominator for recent assignment '${
-                        recentEvent.querySelector(".infotip a[href]")?.href
-                    }' : `,
-                    err
-                );
-            }
-        }
-    }
-
-    setTimeout(getDenominators, 1000);
-}
 
 Logger.debug("Finished loading all.js");
