+++ package.json
@@ -15,6 +15,7 @@
     "url": "https://github.com/aopell/SchoologyPlus"
   },
   "dependencies": {
+    "dompurify": "2.1.1",
     "izitoast": "1.4.0",
     "jquery": "3.3.1",
     "jquery-contextmenu": "2.9.2",
@@ -25,6 +26,7 @@
   "devDependencies": {
     "@trivago/prettier-plugin-sort-imports": "^4.0.0",
     "@types/chrome": "0.0.212",
+    "@types/dompurify": "^2.1.0",
     "@types/jquery": "3.3.1",
     "@types/jquery.contextmenu": "^1.7.38",
     "@types/jqueryui": "1.12.1",
